Extract color output handling for reuse
Move the colorize stuff from the reporter to utils, use it in
VerboseHttp.

What's there now is merely a demonstration of capabilities, to be
resolved by further discussion.
